Create the date field class
The gov uk DfE gem expects on the input value the object of the form
builder to return a day a month and a year.

Unfortunately I can't use the Ruby standard Date class because
there is possibillities that the user can enter invalid dates
(such as '32/13/2020') and we need to render those invalid values
in the input while we show the date invalid error, so
the only alternative is to create a object that responds to
day, month and year.

For more details see:
https://github.com/DFE-Digital/govuk_design_system_formbuilder/blob/master/lib/govuk_design_system_formbuilder/elements/date.rb

On current date the line that the gem called is

value = @builder.object.try(@attribute_name).try(segment)

Hence this Date field object
