+++ app/models/metadata_presenter/page_answers.rb
@@ -1,22 +1,4 @@
 module MetadataPresenter
-  class DateField
-    attr_reader :day, :month, :year
-
-    def initialize(day:, month:, year:)
-      @day = day
-      @month = month
-      @year = year
-    end
-
-    def present?
-      !blank?
-    end
-
-    def blank?
-      @day.blank? || @month.blank? || @year.blank?
-    end
-  end
-
   class PageAnswers
     include ActiveModel::Model
     include ActiveModel::Validations
